+++ app/src/main/java/com/borshevskiy/currencyexchangetestapp/data/network/model/USD.kt
@@ -7,5 +7,5 @@ data class USD(
     @SerializedName("code")
     val code: String,
     @SerializedName("value")
-    val value: Int
+    val value: Double
 )
